+++ packages/transformer/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @itwin/imodel-transformer
 
-This log was last generated on Fri, 18 Aug 2023 23:12:28 GMT and should not be manually modified.
+This log was last generated on Tue, 26 Sep 2023 16:19:57 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.2
+
+Tue, 26 Sep 2023 16:19:57 GMT
+
+### Patches
+
+- Fix aspect queries when class name is reserved SQLite keyword ([commit](https://github.com/iTwin/transformer/commit/dc462020a152694640355f06d6263aae464c52f6))
+
+## 0.4.1
+
+Wed, 20 Sep 2023 15:35:21 GMT
+
+### Patches
+
+- revert to original behavior of provenance ExternalSourceAspect version behavior ([commit](https://github.com/iTwin/transformer/commit/46373c33920c763ba3eb866fc415e433aa0952e6))
+
+## 0.4.0
+
+Mon, 11 Sep 2023 12:37:44 GMT
+
+### Minor changes
+
+- Add detached ElementAspect exporting ([commit](https://github.com/iTwin/transformer/commit/4c404f3980ec7f4e6a3f3a0b746701e4c6f77d92))
 
 ## 0.3.2